Require binary filenames to end in ".bin"
Without any kind of filter for what binaries it supports, this loader
will always appear at the top of the list for all binaries. This got
pretty annoying when I was importing a bunch of non-ASMedia binaries.

To start, let's just filter on the firmware filename. We can come up
with a better heuristic later.